Button Reduction

+++ public/javascript/delete-project.js
@@ -5,14 +5,19 @@ async function deleteFormHandler(event) {
     window.location.toString().split('/').length - 1
   ];
 
-  const response = await fetch(`/api/projects/${id}`, {
-    method: 'DELETE'
-  });
+  const confirmDelete = confirm('Are you sure you would like to delete this project and all tasks?')
+  if (confirmDelete == true) {
 
-  if (response.ok) {
-    document.location.replace('/dashboard');
-  } else {
+    const response = await fetch(`/api/projects/${id}`, {
+      method: 'DELETE'
+    })
+    if (response.ok) {
+      document.location.replace('/dashboard');
+    } else {
     alert(response.statusText);
+    }
+  } else {
+    return;
   }
 };
